requirements.txt
flask
sentence-transformers
torch
pdfplumber
numpy
scikit-learn
python-docx
